// pro/src/apiClient/hey-api/zod.gen.ts
export const zPokeTodoPriority = z.enum(['low', 'medium', 'high'])

/**
 * CreatePokeTodoBodyModel
 */
export const zCreatePokeTodoBodyModel = z.object({
  description: z.optional(z.union([z.string(), z.null()])).default(null),
  dueDate: z.optional(z.union([z.iso.datetime(), z.null()])).default(null),
  priority: z.optional(zPokeTodoPriority).default('low'),
  title: z.string().max(255, { error: 'Title cannot exceed 255 characters' }),
})
// ...
export const zUpdatePokeTodoBodyModel = z.object({
  description: z.optional(z.union([z.string(), z.null()])).default(null),
  dueDate: z.optional(z.union([z.iso.datetime(), z.null()])).default(null),
  isCompleted: z.optional(z.union([z.boolean(), z.null()])).default(null),
  priority: z.optional(z.union([zPokeTodoPriority, z.null()])).default(null),
  title: z
    .optional(
      z.union([
        z.string().max(255, { error: 'Title cannot exceed 255 characters' }),
        z.null(),
      ])
    )
    .default(null),
})